add working day

+++ API/DTO/WorkingDayDTO.cs
@@ -2,7 +2,7 @@
 {
     public class WorkingDayDTO
     {
-        public int NumberOfHours { get; set; }
+        public double NumberOfHours { get; set; }
         public DateTime Date {  get; set; }
         public string WorkStatus { get; set; }
     }
